Moves the extracted form state container helpers into the containers directory
