+++ client/helpers/connect-form.js
